refactor: rename file and folder to EmptyCart
