+++ src/pages/Cart/EmptyCart/index.ts
@@ -1,0 +1,2 @@
+import EmptyCart from './EmptyCart';
+export { EmptyCart };
