--- src/pages/Cart/EmptyMessage/index.ts
@@ -1,2 +1,0 @@
-import EmptyCart from './EmptyMessage';
-export { EmptyCart };
